+++ app/lib/ui/about/detail/about_file_screen.dart
@@ -26,11 +26,7 @@ class AboutFileScreen extends StatefulWidget {
   final String fileMenuItem;
   final bool darkModeEnabled;
 
-  const AboutFileScreen(
-      {super.key,
-      required this.title,
-      required this.fileMenuItem,
-      required this.darkModeEnabled});
+  const AboutFileScreen({super.key, required this.title, required this.fileMenuItem, required this.darkModeEnabled});
 
   @override
   AboutFileScreenState createState() => AboutFileScreenState();
@@ -285,14 +281,12 @@ a:link {
 
         </style>''';
 
-    String cssAndroidThemeStyle =
-        cssAndroid.replaceFirst('</style>', '$themeStyle</style>');
+    String cssAndroidThemeStyle = cssAndroid.replaceFirst('</style>', '$themeStyle</style>');
 
     return convertToHtml(fileContent, cssApple, cssAndroidThemeStyle);
   }
 
-  String convertToHtml(
-      String partialHTMLText, String cssApple, String cssAndroid) {
+  String convertToHtml(String partialHTMLText, String cssApple, String cssAndroid) {
     String head = """
     <!doctype html>
 <meta name='viewport' content='width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no'>